services/public-api/pyproject.toml
[project]
name = "public-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "google-cloud-firestore>=2.26.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.42.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "types-pyyaml>=6.0.12.20250915",
]
